requirements.txt
Django==3.2.9
djangorestframework==3.12.4
rdflib==6.0.2
django-environ==0.8.1